CNN_CPU_Stress/requirements.txt
numpy
pandas
psutil
pillow

torch
torchvision
